src/isaac_sim_validation/requirements.txt
# Requirements for Isaac Sim Validation Framework

# Core scientific libraries
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization libraries
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data processing and file formats
pyyaml>=5.4.0
jsonschema>=3.2.0
h5py>=3.7.0

# Robotics and simulation (when available)
# These are optional and may not be installable in all environments
# ros-nodes # Placeholder for ROS 2 Python bindings
# isaac-sim # Placeholder for Isaac Sim Python API

# Performance and optimization
scikit-optimize>=0.9.0
optuna>=2.10.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pathlib>=1.0.0